Rename packages That way we no longer have any strong link with the xolstice code anymore (avoid confusion)
