--- src/main/java/org/xolstice/maven/toolchain/protobuf/ProtobufToolchain.java
@@ -1,31 +1,0 @@
-package org.xolstice.maven.toolchain.protobuf;
-
-/*
- * Copyright (c) 2016 Maven Protocol Buffers Plugin Authors. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import org.apache.maven.toolchain.Toolchain;
-
-/**
- * A tool chain for protobuf compiler (protoc).
- *
- * @since 0.2.0
- */
-public interface ProtobufToolchain extends Toolchain {
-
-    String getProtocExecutable();
-
-    void setProtocExecutable(String protocExecutable);
-}
